debugging a problem with docid > documents that turns out to be caused by the indexer in index.c when a blank file is indexed.
git-svn-id: https://groke.otago.ac.nz/ant/ant@538 7489a7b4-4588-4892-8acd-774bacfab51c

+++ source/parser.c
@@ -149,7 +149,7 @@ else if (*current & 0x80)		// UTF-8 character
 			{
 			while (ischinese(current))		// don't need to check for '\0' because that isn't a Chinese character
 				current += utf8_bytes(current);
-			segment(start, current - start);
+			segment(start, (long)(current - start));
 			return get_next_token();
 			}
 		else
