+++ source/search_engine_result.h
@@ -88,10 +88,13 @@ public:
 				min_in_top_k = accumulator_pointers[current]->get_rsv();
 		}
 
-	template <class T> inline void add_rsv(size_t index, T score)
+	template <class T> inline void add_rsv(long index, T score)
 		{
 		ANT_search_engine_accumulator::ANT_accumulator_t was;
 		ANT_search_engine_accumulator *which = accumulator + index; 
+
+//		if (index > documents)
+//			printf("docid %d too big\n", index);
 
 		if ((was = which->get_rsv()) <= min_in_top_k)
 			{									// were' not in the top-k but we might end up there.
